+++ NWaves.Tests/SignalTests/TestDiscreteSignal.cs
@@ -8,19 +8,19 @@ namespace NWaves.Tests.SignalTests
     public class TestDiscreteSignal
     {
         private readonly DiscreteSignal _signal = 
-            new DiscreteSignal(8000, new double[] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 });
+            new DiscreteSignal(8000, new float[] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 });
 
         private readonly DiscreteSignal _constant = 
-            new DiscreteSignal(8000, new double[] { 5, 5, 5, 5, 5 });
+            new DiscreteSignal(8000, new float[] { 5, 5, 5, 5, 5 });
 
         private readonly DiscreteSignal _small = 
-            new DiscreteSignal(16000, new double[] { 5, 2, 4 });
+            new DiscreteSignal(16000, new float[] { 5, 2, 4 });
 
         [Test]
         public void TestInitializeWithBadSamplingRate()
         {
-            Assert.Throws<ArgumentException>(() => { var s = new ComplexDiscreteSignal(0, new double[] { 1 }); });
-            Assert.Throws<ArgumentException>(() => { var s = new ComplexDiscreteSignal(-8000, new double[] { 1 }); });
+            Assert.Throws<ArgumentException>(() => { var s = new ComplexDiscreteSignal(0, new float[] { 1 }); });
+            Assert.Throws<ArgumentException>(() => { var s = new ComplexDiscreteSignal(-8000, new float[] { 1 }); });
         }
 
         [Test]
@@ -28,7 +28,7 @@ namespace NWaves.Tests.SignalTests
         {
             Assert.Throws<ArgumentException>(() =>
             {
-                var s = new ComplexDiscreteSignal(8000, new double[] { 1, 2 }, new double[] { 3 });
+                var s = new ComplexDiscreteSignal(8000, new float[] { 1, 2 }, new float[] { 3 });
             });
         }
 
@@ -39,7 +39,7 @@ namespace NWaves.Tests.SignalTests
             var delayed = _signal.Delay(3);
 
             //Assert
-            Assert.That(delayed.Samples, Is.EqualTo(new double[] { 0, 0, 0, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 }));
+            Assert.That(delayed.Samples, Is.EqualTo(new float[] { 0, 0, 0, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 }));
         }
 
         [Test]
@@ -49,7 +49,7 @@ namespace NWaves.Tests.SignalTests
             var delayed = _signal.Delay(-3);
 
             //Assert
-            Assert.That(delayed.Samples, Is.EqualTo(new double[] { 3, 4, 5, 6, 7, 8, 9 }));
+            Assert.That(delayed.Samples, Is.EqualTo(new float[] { 3, 4, 5, 6, 7, 8, 9 }));
         }
 
         [Test]
@@ -65,7 +65,7 @@ namespace NWaves.Tests.SignalTests
             var combination = _signal.Superimpose(_constant);
 
             //Assert
-            Assert.That(combination.Samples, Is.EqualTo(new double[] { 5, 6, 7, 8, 9, 5, 6, 7, 8, 9 }));
+            Assert.That(combination.Samples, Is.EqualTo(new float[] { 5, 6, 7, 8, 9, 5, 6, 7, 8, 9 }));
         }
 
         [Test]
@@ -75,7 +75,7 @@ namespace NWaves.Tests.SignalTests
             var combination = _constant.Superimpose(_signal);
 
             //Assert
-            Assert.That(combination.Samples, Is.EqualTo(new double[] { 5, 6, 7, 8, 9, 5, 6, 7, 8, 9 }));
+            Assert.That(combination.Samples, Is.EqualTo(new float[] { 5, 6, 7, 8, 9, 5, 6, 7, 8, 9 }));
         }
 
         [Test]
@@ -92,8 +92,8 @@ namespace NWaves.Tests.SignalTests
             var concatenation2 = _constant + _signal;
 
             //Assert
-            Assert.That(concatenation1.Samples, Is.EqualTo(new double[] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 5, 5, 5, 5, 5 }));
-            Assert.That(concatenation2.Samples, Is.EqualTo(new double[] { 5, 5, 5, 5, 5, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 }));
+            Assert.That(concatenation1.Samples, Is.EqualTo(new float[] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 5, 5, 5, 5, 5 }));
+            Assert.That(concatenation2.Samples, Is.EqualTo(new float[] { 5, 5, 5, 5, 5, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 }));
         }
 
         [Test]
@@ -110,8 +110,8 @@ namespace NWaves.Tests.SignalTests
             var repeated2 = _small * 3;
 
             //Assert
-            Assert.That(repeated1.Samples, Is.EqualTo(new double[] { 5, 2, 4, 5, 2, 4, 5, 2, 4 }));
-            Assert.That(repeated2.Samples, Is.EqualTo(new double[] { 5, 2, 4, 5, 2, 4, 5, 2, 4 }));
+            Assert.That(repeated1.Samples, Is.EqualTo(new float[] { 5, 2, 4, 5, 2, 4, 5, 2, 4 }));
+            Assert.That(repeated2.Samples, Is.EqualTo(new float[] { 5, 2, 4, 5, 2, 4, 5, 2, 4 }));
         }
 
         [Test]
@@ -133,7 +133,7 @@ namespace NWaves.Tests.SignalTests
             var slice = _signal[3, 7];
 
             // Assert
-            Assert.That(slice.Samples, Is.EqualTo(new double[] { 3, 4, 5, 6 }));
+            Assert.That(slice.Samples, Is.EqualTo(new float[] { 3, 4, 5, 6 }));
         }
 
         [Test]
@@ -156,7 +156,7 @@ namespace NWaves.Tests.SignalTests
             var first = _signal.First(3);
 
             // Assert
-            Assert.That(first.Samples, Is.EqualTo(new double[] { 0, 1, 2 }));
+            Assert.That(first.Samples, Is.EqualTo(new float[] { 0, 1, 2 }));
         }
 
         [Test]
@@ -166,7 +166,7 @@ namespace NWaves.Tests.SignalTests
             var last = _signal.Last(3);
 
             // Assert
-            Assert.That(last.Samples, Is.EqualTo(new double[] { 7, 8, 9 }));
+            Assert.That(last.Samples, Is.EqualTo(new float[] { 7, 8, 9 }));
         }
 
         [Test]
